refactor: move profile specific code into dedicated files
