+++ src/profiles/prof_ch_iec_30168_basic_passcode.c
@@ -1,0 +1,20 @@
+/* SPDX-License-Identifier: MPL-2.0 */
+/**********************************************************************
+ * Copyright (c) 2025, Siemens AG
+ **********************************************************************/
+
+#include <gta_api/gta_api.h>
+#include "../gta_sw_provider.h"
+
+GTA_SWP_DEFINE_FUNCTION(bool, context_open,
+(
+    struct gta_sw_provider_context_params_t * p_context_params,
+    gta_errinfo_t * p_errinfo
+))
+{
+    return true;
+}
+
+const struct profile_function_list_t fl_prof_ch_iec_30168_basic_passcode = {
+    .context_open = context_open,
+};
